test: add test for client page snapshot
